Move this into Object Builders API
