+++ fabric-object-builder-api-v1/src/main/java/net/fabricmc/fabric/mixin/object/builder/BlockStateProviderTypeInvoker.java
@@ -1,4 +1,4 @@
-package net.fabricmc.fabric.mixin.tree;
+package net.fabricmc.fabric.mixin.object.builder;
 
 import com.mojang.serialization.Codec;
 import net.minecraft.world.gen.stateprovider.BlockStateProvider;
