--- fabric-tree-api-v1/src/main/resources/fabric-tree-api-v1.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.tree",
-  "compatibilityLevel": "JAVA_16",
-  "mixins": [
-    "BlockStateProviderTypeInvoker",
-    "FoliagePlacerTypeInvoker",
-    "TreeDecoratorTypeInvoker",
-    "TrunkPlacerTypeInvoker"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
