--- fabric-tree-api-v1/src/main/resources/fabric.mod.json
@@ -1,34 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-tree-api-v1",
-  "name": "Fabric Tree API (v1)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-tree-api-v1/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.4.0",
-    "fabric-api-base": "*",
-    "fabric-tag-extensions-v0": "*"
-  },
-  "entrypoints": {
-    "main": [
-    ]
-  },
-  "description": "Tree creation utilities",
-  "mixins": [
-    "fabric-tree-api-v1.mixins.json"
-  ],
-  "custom": {
-    "fabric-api:module-lifecycle": "stable"
-  }
-}
