+++ settings.gradle
@@ -55,4 +55,3 @@ include 'fabric-structure-api-v1'
 include 'fabric-tag-extensions-v0'
 include 'fabric-textures-v0'
 include 'fabric-tool-attribute-api-v1'
-include 'fabric-tree-api-v1'
